feat : add truncate text configuration
Allows the definition of isTruncated and fieldSize on the receipt form
map (configuration file).
When isTruncated is set to true, multi-line fields are truncated to the
maximum lenght defined by fieldSize.
It determines the occupied lenght by measuring the size of each
character of the text based on font and weight.